make generic `fix_state_vars` method
